requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.11.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-guardian==2.4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.1.0